+++ core/src/commonTest/kotlin/nl/adaptivity/xmlutil/TestBufferedXmlReader.kt
@@ -21,7 +21,9 @@
 package nl.adaptivity.xmlutil
 
 import nl.adaptivity.xmlutil.core.KtXmlReader
+import nl.adaptivity.xmlutil.core.KtXmlWriter
 import nl.adaptivity.xmlutil.core.impl.multiplatform.StringReader
+import nl.adaptivity.xmlutil.core.impl.multiplatform.StringWriter
 import kotlin.test.Ignore
 import kotlin.test.Test
 
@@ -69,7 +71,12 @@ class TestBufferedXmlReader : TestCommonReader() {
 
     @Test
     fun testProcessingInstruction() {
-        testProcessingInstruction(::createReader)
+        testProcessingInstruction(::createReader) { KtXmlWriter(StringWriter())}
+    }
+
+    @Test
+    fun testProcessingInstructionDom() {
+        testProcessingInstruction(::createReader) { DomWriter() }
     }